mesa-asahi: get working for the 24.11 channel
If you're like me and:
1) don't deploy using a flake
2) use the stable 24.11 channel instead of unstable

you will find that the oliverbestmann /tpwrules fixes for mesa-asahi don't work due to the de-vendoring.  24.11 doesn't have the libgbm separation so we need to keep the vendoring.

These changes apply the appropriate changes from the mesa in nixpkgs so that we can build mesa>=24.3.  If you are using nixos-unstable, don't use this commit -- follow along with how tpwrules is maintaining mesa.